r2engine: first pass at animated models. Got microbat to flap his wings a bit. Still can't add more animations

+++ r2engine/src/r2/Core/Engine.cpp
@@ -279,7 +279,7 @@ namespace r2
         return 0;
     }
     
-    u32 Engine::GetTicks() const
+    f64 Engine::GetTicks() const
     {
         if(mGetTicksFunc)
             return mGetTicksFunc();
